Expand apiUtils, refactor tests and group them by methods, add info to readme
